requirements.txt
torch
torchvision
opencv-python-headless
numpy
scikit-learn
matplotlib
pillow
